deps: Update to Typst v0.15.0
Unfortunately, the file handling business in typst-kit is not quite as
extensible as I wished. Which is a shame because I had the opportunity
to review it, but not enough time and foresight to see the issues I
found while trying to use it here.

Because of these difficulties this patch doesn't use the new file store
API, still replicating its inner code for now. This should be resolved
in a future patch by either making it more extensible, or at least
exposing its internals (such as BOM removal).

TODO: Add changelog.

+++ crates/tytanic-core/src/config.rs
@@ -82,7 +82,7 @@ pub struct ProjectDefaults {
     ///
     /// Defaults to `144.0`.
     #[serde(default = "default_ppi")]
-    pub ppi: f32,
+    pub ppi: f64,
 
     /// The default maximum allowed delta per pixel.
     ///
@@ -112,7 +112,7 @@ fn default_direction() -> Direction {
     Direction::Ltr
 }
 
-fn default_ppi() -> f32 {
+fn default_ppi() -> f64 {
     144.0
 }
